requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bidict==0.22.1
billiard==4.1.0
bleach==6.1.0
Brotli==1.1.0
CacheControl==0.13.1
cached-property==1.5.2
cachetools==5.3.2
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.36.0
constantly==15.1.0
cryptography==41.0.5
daphne==4.0.0
dj-database-url==2.1.0
dj-url-filter==0.4.4
Django==4.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-genericrelationview @ git+https://github.com/macornwell/genericrelationview/@e2ba42016709492b336a7f0b11764192e1f70654
django-grappelli @ git+https://github.com/sehmaschine/django-grappelli.git@c676de89ddcd6a4ef0a21c2553d480e333f0c68d
django-nested-admin==4.0.2
django-nested-inline==0.4.6
django-restql==0.15.3
djangorestframework==3.14.0
djira @ git+https://github.com/oasisMystre/djira@9747f10181dda6defc139641a36285cf8e9452fa
drf-writable-nested==0.7.0
firebase-admin==6.2.0
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.8.3
inapppy==2.5.2
incremental==22.10.0
kombu==5.3.2
Markdown==3.3.7
martor @ git+https://github.com/oasisMystre/django-markdown-editor-grappelli@a3f7639434cd3a9d41d4bf6e21ae885a0e903ef9
msgpack==1.0.7
multidict==6.0.4
mysqlclient==2.2.0
ngrok==0.10.1
numpy==1.26.1
oauth2client==4.1.3
openai==0.28.1
Pillow==10.0.1
priority==1.3.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
psycopg==3.1.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyPEG2==2.15.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-monkey-business==1.0.0
python-socketio==5.10.0
pytz==2023.3.post1
rave-python @ git+https://github.com/oasisMystre/Python-v2@9ad9e334f4bde5181e6ab98da2da4d69026abbdc
requests==2.31.0
rest-framework-generic-relations==2.1.0
rsa==4.9
sentry-sdk==1.32.0
service-identity==23.1.0
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.8
webencodings==0.5.1
whitenoise==6.6.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
